add move unit tests for config and add mock utility
